refactor: rename @srcbook/cells to @srcbook/components
