+++ packages/components/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@srcbook/cells",
+  "name": "@srcbook/components",
   "version": "0.0.1",
   "type": "module",
   "main": "./dist/index.js",
