+++ packages/web/package.json
@@ -34,7 +34,7 @@
     "@radix-ui/react-tabs": "^1.1.0",
     "@radix-ui/react-tooltip": "^1.1.2",
     "@srcbook/shared": "workspace:^",
-    "@srcbook/cells": "workspace:^",
+    "@srcbook/components": "workspace:^",
     "@uiw/codemirror-themes": "^4.23.2",
     "@uiw/react-codemirror": "^4.23.2",
     "class-variance-authority": "^0.7.0",
