+++ packages/web/tailwind.config.js
@@ -7,7 +7,7 @@ module.exports = {
   content: [
     './index.html',
     './src/**/*.{css,ts,tsx}',
-    path.join(path.dirname(require.resolve('@srcbook/cells')), '**/*.{css,js,ts,tsx}'),
+    path.join(path.dirname(require.resolve('@srcbook/components')), '**/*.{css,js,ts,tsx}'),
   ],
   prefix: '',
   theme: {
